Adding support for Azure Functions
